feat: New script that automatically can generate all the folder images (#95)
* chore: Rearrange project folder structure

- Moved base images to the icons directory
- Created a separate fonts directory for font files
- Updated references and paths throughout the codebase
- Deleted outdated folder images in the output directory
- Cleaned up unused files to ensure up-to-date icons

* feat: Add base images for Windows 11 and macOS Ventura

This commit adds the base folder images for Windows 11 and macOS Ventura. The images are located in the respective folders under the 'src/icons' directory.

The base images are necessary for generating the custom folder icons for each operating system.

* feat: Add Font Awesome repository as submodule for SVG masks

This commit adds the Font Awesome repository as a submodule to the project. The submodule is located in the 'masks/fontawesome' directory.

The Font Awesome repository will be used to access SVG icons, which will be used as masks for the custom folder icons.

* refactor: Add utils.py and add environment variable handling

This commit refactors the utils.py file to improve code organization and readability.
It introduces environment variable handling for the `MASK_FOLDER_PATH`, `BASE_FOLDER_PATH`, `IMAGE_FILE_EXTENSION`, and `MASK_FILE_EXTENSION` variables. The `load_image_files` function now accepts an optional `image_file_extension` parameter, allowing flexibility in specifying the file extension for loading image files.
Additionally, the commit adds error handling to check for missing environment variables and raises `ValueError` if any of them are not set.
The main section of the script has been updated to demonstrate the functionality by loading mask files and base files and printing their paths.

* chore: remove Font-Awesome gitsubmodule

* chore: remove Font-Awesome gitsubmodule

* refactor: add env variable and requirements.txt

* refactor: add env file and requirements

* refactor: restructured image folders

* feat: update generator and utils

* chore: update gitignore

* chore: remove the cache file from repo

* chore: update gitignore with pyc

* chore: update gitignore with python artifacts

* test: add unit tests

* refactor: fix package issue

* bugfix: fix filename issue in generator

* chore: change image folder structure

* fix: fix output folder grouping

* feat: add image callibration

* feat: update function to support image calibration

* fix: fix typo in calibration file

* chore: removed unused variable

* chore: remove print commands

* fix: update alpha value

* style: fix code styling

* refactor: removed commented code

* fix: fix pytest path issue

* test: update unit tests

* chore: update requirements.txt

* docs: update utils.py python doc

* docs: update python doc for generator.py

* docs: increase main.py readability